Trait pallet_tidefi_rpc::TidefiApiServer

pub trait TidefiApiServer<BlockHash, AccountId>: Sized + Send + Sync + 'static {
    fn get_assets(
        &self,
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(CurrencyId, CurrencyMetadata<Vec<u8>>)>>;
    fn get_account_balance(
        &self,
        account_id: AccountId,
        asset_id: CurrencyId,
        at: Option<BlockHash>
    ) -> RpcResult<CurrencyBalance<BalanceInfo>>;
    fn get_account_balances(
        &self,
        account_id: AccountId,
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(CurrencyId, CurrencyBalance<BalanceInfo>)>>;
    fn get_account_stakes(
        &self,
        account_id: AccountId,
        at: Option<BlockHash>
    ) -> RpcResult<Vec<(CurrencyId, Stake<BalanceInfo, BlockNumber>)>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        AccountId: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the TidefiApi RPC API.

Required Methods

fn get_assets(
&self,
at: Option<BlockHash>
) -> RpcResult<Vec<(CurrencyId, CurrencyMetadata<Vec<u8>>)>>

fn get_account_balance(
&self,
account_id: AccountId,
asset_id: CurrencyId,
at: Option<BlockHash>
) -> RpcResult<CurrencyBalance<BalanceInfo>>

fn get_account_balances(
&self,
account_id: AccountId,
at: Option<BlockHash>
) -> RpcResult<Vec<(CurrencyId, CurrencyBalance<BalanceInfo>)>>

fn get_account_stakes(
&self,
account_id: AccountId,
at: Option<BlockHash>
) -> RpcResult<Vec<(CurrencyId, Stake<BalanceInfo, BlockNumber>)>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
BlockHash: Send + Sync + 'static + DeserializeOwned,
AccountId: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.